+++ mmdet/core/bbox/iou_calculators/__init__.py
@@ -1,4 +1,4 @@
 from .builder import build_iou_calculator
-from .iou2d_calculator import BboxOverlaps2D
+from .iou2d_calculator import BboxOverlaps2D, bbox_overlaps
 
-__all__ = ['build_iou_calculator', 'BboxOverlaps2D']
+__all__ = ['build_iou_calculator', 'BboxOverlaps2D', 'bbox_overlaps']
